+++ src/__tests__/resolveErrorMessage.test.js
@@ -1,6 +1,7 @@
 import resolveErrorMessage from '../resolveErrorMessage';
 import HttpErrorMessageConfig from '../entity/HttpErrorMessageConfig';
 import ErrorMessageConfig from '../entity/ErrorMessageConfig';
+import ErrorMessage from '../entity/ErrorMessage';
 
 describe('resolveErrorMessage', () => {
     test.each([
@@ -8,7 +9,7 @@ describe('resolveErrorMessage', () => {
             'Not http error',
             null,
             new ErrorMessageConfig(new HttpErrorMessageConfig('Default message'), 'Error Default message'),
-            'Error Default message',
+            new ErrorMessage('Error Default message'),
         ],
         [
             'Resolve custom message',
@@ -21,7 +22,23 @@ describe('resolveErrorMessage', () => {
 
                     return null;
                 }),
-            'Some error',
+            new ErrorMessage('Some error'),
+        ],
+        [
+            'Http error with correlation key in response header',
+            {
+                config: {},
+                request: {},
+                response: {
+                    status: 500,
+                    data: {},
+                    headers: {
+                        'paysera-correlation-id': 'abc',
+                    },
+                },
+            },
+            new ErrorMessageConfig(new HttpErrorMessageConfig('Default message'), 'Error Default message'),
+            new ErrorMessage('Default message', ['Error code: abc']),
         ],
     ])(
         '%s',