+++ src/index.js
@@ -5,5 +5,6 @@ export { default as ErrorMessageConfig } from './entity/ErrorMessageConfig';
 export { default as ErrorDataMessageConfig } from './entity/ErrorDataMessageConfig';
 export { default as HttpErrorMessageConfig } from './entity/HttpErrorMessageConfig';
 export { default as ErrorStatusMessageConfig } from './entity/ErrorStatusMessageConfig';
+export { default as ErrorMessage } from './entity/ErrorMessage';
 
 export { default as resolveErrorMessage } from './resolveErrorMessage';